refactor code to testutil.exercise
