+++ src/test/java/seedu/zerotoone/model/exercise/ExerciseListTest.java
@@ -24,7 +24,7 @@ import seedu.zerotoone.model.exercise.Exercise;
 import seedu.zerotoone.model.exercise.ExerciseList;
 import seedu.zerotoone.model.exercise.ReadOnlyExerciseList;
 import seedu.zerotoone.model.exercise.exceptions.DuplicateExerciseException;
-import seedu.zerotoone.testutil.ExerciseBuilder;
+import seedu.zerotoone.testutil.exercise.ExerciseBuilder;
 
 public class ExerciseListTest {
